asdfasdgsdgsdgfsdfgsadfasdf

+++ api/src/app/Controllers/ProductsController.js
@@ -1,8 +1,9 @@
 const { PrismaClient } = require('@prisma/client')
 
 const prisma = new PrismaClient()
-const verifyImage = (imagem) =>{
-    const re = /(http[s]*:\/\/)([a-z\-_0-9\/.]+)\.([a-z.]{2,3})\/([a-z0-9\-_\/._~:?#\[\]@!$&'()*+,;=%]*)([a-z0-9]+\.)(jpg|jpeg|png)/i
+const verifyImage = (imagem) => {
+    const re =
+        /(http[s]*:\/\/)([a-z\-_0-9\/.]+)\.([a-z.]{2,3})\/([a-z0-9\-_\/._~:?#\[\]@!$&'()*+,;=%]*)([a-z0-9]+\.)(jpg|jpeg|png)/i
     return re.test(imagem)
 }
 class ProductsController {
@@ -24,7 +25,13 @@ class ProductsController {
             await prisma.$connect()
 
             const { nome, imagem, preco } = req.body
-            if (nome == '' || imagem == '' || preco == '' || !verifyImage(imagem) || preco<0) {
+            if (
+                nome == '' ||
+                imagem == '' ||
+                preco == '' ||
+                !verifyImage(imagem) ||
+                preco < 0
+            ) {
                 return res.status(406).json({ message: 'Invalid data' })
             } else {
                 const createdProduct = await prisma.produtos.create({
@@ -53,15 +60,20 @@ class ProductsController {
 
             const { id } = req.body
             if (id != '') {
+                const foiEmprestado = await prisma.emprestimos.findMany({
+                    where: { idProduto: id.toString() },
+                })
+                if (foiEmprestado.length >= 1) {
+                    return res.status(403).json('Fez empréstimo')
+                }
+
                 const deletedProduct = await prisma.produtos.delete({
                     where: { id: id },
                 })
-                return res
-                    .status(200)
-                    .json({
-                        message: 'Product deleted with sucess',
-                        deletedProduct,
-                    })
+                return res.status(200).json({
+                    message: 'Product deleted with sucess',
+                    deletedProduct,
+                })
             }
         } catch (e) {
             return res.status(500).json(e)
